Log sendmail output
There's really no way to debug a broken sendmail setup when the output
is thrown away. Stdout and stderr were passed to the exception as
attributes but do not appear to be used there.

I don't believe there's any other good way to merge stdout/stderr and
preserve the sequential order than to pass one to the other and the
other to PIPE.

Alternatively to _Log.debug, we could pass output to _LOG.error only if
there's a non-zero exit code, but I'm biased towards maximum verbosity
at the debug log level.